+++ lib/ui/explore/ExploreDiningDetailPanel.dart
@@ -24,8 +24,9 @@ import 'package:illinois/ext/Dining.dart';
 import 'package:illinois/ext/Explore.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/model/RecentItem.dart';
+import 'package:rokwire_plugin/rokwire_plugin.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
-import 'package:illinois/service/DiningService.dart';
+import 'package:illinois/service/Dinings.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/widgets/FilterWidgets.dart';
@@ -107,7 +108,7 @@ class _DiningDetailPanelState extends State<ExploreDiningDetailPanel> implements
     if(!dining!.hasDiningSchedules){
       _isDiningLoading = true;
 
-      DiningService().loadBackendDinings(false, null, _locationData).then((List<Dining>? dinings){
+      Dinings().loadBackendDinings(false, null, _locationData).then((List<Dining>? dinings){
         if(dinings != null){
           Dining? foundDining = Dining.entryInList(dinings, id: dining!.id);
           if(foundDining != null){
@@ -599,7 +600,7 @@ class _DiningDetailPanelState extends State<ExploreDiningDetailPanel> implements
     if (StringUtils.isEmpty(deepLink)) {
       return;
     }
-    bool? appLaunched = await NativeCommunicator().launchApp({"deep_link": deepLink});
+    bool? appLaunched = await RokwirePlugin.launchApp({"deep_link": deepLink});
     if (appLaunched != true) {
       String storeUrl = orderOnlineDetails!['store_url'];
       url_launcher.launch(storeUrl);
@@ -724,7 +725,7 @@ class _DiningDetailState extends State<_DiningDetail> implements NotificationsLi
   }
 
   void _loadOffers(){
-    DiningService().loadDiningSpecials().then((List<DiningSpecial>? offers){
+    Dinings().loadDiningSpecials().then((List<DiningSpecial>? offers){
       if(offers != null && offers.isNotEmpty){
         _specials = offers.where((entry)=>entry.locationIds!.contains(widget.dining!.id)).toList();
         setState((){});
@@ -736,7 +737,7 @@ class _DiningDetailState extends State<_DiningDetail> implements NotificationsLi
     if(hasMenuData) {
       _isLoading = true;
       DateTime? filterDate = _filterDates![_selectedDateFilterIndex];
-      DiningService().loadMenuItemsForDate(widget.dining!.id, filterDate).then((
+      Dinings().loadMenuItemsForDate(widget.dining!.id, filterDate).then((
           List<DiningProductItem>? items) {
         _productItems = items;
         _productItemsMapping = Map<String, DiningProductItem>();