+++ plugin/lib/service/config.dart
@@ -424,7 +424,7 @@ class Config with Service, NetworkAuthProvider, NotificationsListener {
     return _configEnvironment;
   }
 
-  static ConfigEnvironment get defaultConfigEnvironment {
+  ConfigEnvironment get defaultConfigEnvironment {
     return kReleaseMode ? ConfigEnvironment.production : ConfigEnvironment.dev;
   }
 
@@ -476,6 +476,7 @@ class Config with Service, NetworkAuthProvider, NotificationsListener {
   String? get quickPollsUrl    => JsonUtils.stringValue(platformBuildingBlocks["polls_url"]);
   String? get eventsUrl        => JsonUtils.stringValue(platformBuildingBlocks['events_url']);
   String? get groupsUrl        => JsonUtils.stringValue(platformBuildingBlocks["groups_url"]);
+  String? get contentUrl       => JsonUtils.stringValue(platformBuildingBlocks["content_url"]);
 
   // Getters: otherUniversityServices
   String? get assetsUrl => JsonUtils.stringValue(otherUniversityServices['assets_url']);